requirements.txt
praw
supabase
python-dotenv
discord.py